Extracting loop preprocessing
